Polygen: Fixed some issues with ThreadManager storing references to threads

+++ portfolio/polygen/javascript/threadManager.js
@@ -1,7 +1,7 @@
 // Manages web workers that run in the background.
 export class Thread {
 	static supported = true;
-	static openWorkers = [];
+	static openWorkers = {};
 
 	constructor() {
 		this.callback = null;
@@ -11,20 +11,24 @@ export class Thread {
 
 	open(workerName) {
 		if (!Thread.supported || !window.Worker) {
-			console.warn("WebWorkers aren't supported. Features disabled: color line finding.");
+			console.warn("WebWorkers aren't supported. Some features may not function properly.");
 			Thread.supported = false;
 			return;
 		}
 
-		if (Thread.openWorkers.includes(workerName)) throw new Error("Thread Error: attempting to open a worker that is already open");
+		if (!!Thread.openWorkers[workerName]) throw new Error("Thread Error: attempting to open a worker that is already open");
 
 		this.worker = new Worker(`./javascript/${workerName}.js`);
 		this.workerName = workerName;
 
-		Thread.openWorkers.push(workerName);
+		Thread.openWorkers[workerName] = { worker: workerName, running: false };
 
 		this.worker.onmessage = (data) => {
-			this.callback(data.data);
+			if (!!data.data?._threadStatus) {
+				Thread.openWorkers[this.workerName].running = data.data._threadStatus.running;
+			} else {
+				this.callback(data.data);
+			}
 		};
 
 		this.worker.onerror = (e) => {
@@ -40,8 +44,12 @@ export class Thread {
 		this.callback = callback;
 	}
 
-	close() {
-		this.worker.terminate();
-		Thread.openWorkers.splice(Thread.openWorkers.indexOf(this.workerName), 1);
+	close(force = false) {
+		if (force || !Thread.openWorkers[this.workerName].running) {
+			this.worker.terminate();
+			delete Thread.openWorkers[this.workerName];
+		} else {
+			throw new Error("Thread Error: attempted to close a worker that is running. To do this, use the 'force' parameter")
+		}
 	}
 }
